(chore) reorganized the components.redone the logn
       Reasons
=====================
1. Components are too disorganized.
2. Lgoin/registration is not real-time.

       Changes
=====================
1. Components reorgd.
2. Login is now handled in one-page app.
