+++ frontend/eureka/src/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule }             from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
  
-import { WelcomeScreenComponent }   from './welcome-screen.component';
-import { MyTestComponent } from './my-test.component'
+import { WelcomeScreenComponent } from './components/welcome-screen/welcome-screen.component';
+import { MyTestComponent } from './components/my-test/my-test.component'
 
 const routes: Routes = [
   { path: '', redirectTo: '/welcome', pathMatch: 'full' },
