+++ frontend/eureka/src/app/app/app.component.ts
