+++ frontend/eureka/src/app/components/register-form/register-form.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { UserService } from './user.service';
+import { UserService } from '../../services/user.service';
 
 @Component({
   selector: 'register-form',
