+++ frontend/eureka/src/app/components/welcome-screen/welcome-screen.component.html
@@ -4,9 +4,20 @@
             <div class="row">
                 <div class="col-md-8">
                     <div class="jumbotron">
-                        <h2>
-                            {{message}}
-                        </h2>
+
+                        <div *ngIf="! user.name">
+                                <h2>
+                                    Welcome stranger!
+                                </h2>
+                            </div>
+
+                        <div *ngIf="user.name">
+                            <h2>
+                                    Welcome, {{user.name}}
+                                    Welcome, {{user.token}}
+                                </h2>
+                        </div>
+
                         <p>
                             This is a template for a simple marketing or informational website. It includes a large callout called the hero unit and
                             three supporting pieces of content. Use it as a starting point to create something more unique.
